FIXED: seperate utils.py and decorator.py, reason:class and def error
